+++ src/nexus/services/memory/state.py
@@ -18,9 +18,9 @@ from datetime import UTC, datetime
 from typing import Any
 
 from nexus.contracts.types import OperationContext, Permission
-from nexus.core.temporal import parse_datetime
 from nexus.rebac.memory_permission_enforcer import MemoryPermissionEnforcer
 from nexus.services.memory.memory_router import MemoryViewRouter
+from nexus.services.memory.temporal import parse_datetime
 
 logger = logging.getLogger(__name__)
